requirements.txt
pandas
pandas-gbq
google-cloud-bigquery
psycopg2
